back 게시글 삭제시 본인권한확인

+++ src/main/java/com/example/prj2be/controller/board/BoardController.java
@@ -56,8 +56,16 @@ public class BoardController {
    }
 
    @DeleteMapping("remove/{id}")
-   public ResponseEntity remove(@PathVariable Integer id) {
+   public ResponseEntity remove(
+      @PathVariable Integer id,
+      @SessionAttribute(value = "login", required = false) Member login) {
 
+      if (login == null) {
+         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401 에러(비로그인상태)
+      }
+      if (service.hasAccess(id,login)) {
+         return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403 에러(본인인지 모름)
+      }
       if (service.remove(id)) {
          return ResponseEntity.ok().build();
       } else {
